Fix Map Errors
The player no longer appears below the Celadon Mansion stairs and the left piece of the Power Plant exit mat.

The bushes near the Battle Tower are now solid.

Made edit to the ground behind some of Goldenrod City's square roofs. Not enough tiles left in tileset to fix all of them.