Port FRLG+'s Better Softlock Prevention
Instead of using HGSS's check to see if the player is stuck in Cianwood, uses FRLG+'s more robust check instead. HGSS's conditions for the check could be made impossible to satisfy despite being trapped, so it needed to be changed.

The game now checks if the player has the Surf HM, a Pokemon that knows Surf, or a Pokemon that can learn Surf. If the combo of these checks results in the player not being able to get a mon that can use Surf, the NPC will give out a Tentacool. If the player doesn't have the optional Surf HM, this Tentacool will know Surf.

This safeguard is also present on Cinnabar Island, unlike HGSS.

Also used the MON_GIVEN_TO_PARTY,  MON_GIVEN_TO_PC, and MON_CANT_GIVE constants in other scripts that use givemon because I was here.