Fix `WaitForVBlank()`
`WaitForVBlank()` was changed to the bad one in FRLG and Emerald to keep the Wireless Adapter from desyncing and disconnecting link sessions.

Now, the bad one is only used when the Wireless Adapter is communicating, while in normal gameplay the good one is used.

Because Sierra used the function call instead of emitting the direct asm, I believe that this is actually faster than what CD had here before despite the added conditional.